refactor tests and use nose
