+++ setup.cfg
@@ -4,3 +4,7 @@ license_files = LICENSE.txt
 [bdist_wheel]
 # Support for both Python 2 and 3
 universal=1
+
+[nosetests]
+verbosity=3
+detailed-errors=1
